feat: eliminate vue non props warning
