+++ packages/core/src/index.ts
@@ -1,33 +1,4 @@
-import path, { isAbsolute } from 'path';
-import fs from 'fs';
-import { dirname } from 'path';
 import { Editor } from './shared/constant';
-import { startServer, enhanceCode, normalizePath, parseSFC } from './server';
-
-let compatibleDirname = '';
-
-function fileURLToPath(fileURL: string) {
-  let filePath = fileURL;
-  if (process.platform === 'win32') {
-    filePath = filePath.replace(/^file:\/\/\//, '');
-    filePath = decodeURIComponent(filePath);
-    filePath = filePath.replace(/\//g, '\\');
-  } else {
-    filePath = filePath.replace(/^file:\/\//, '');
-    filePath = decodeURIComponent(filePath);
-  }
-  return filePath;
-}
-
-if (typeof __dirname !== 'undefined') {
-  compatibleDirname = __dirname;
-} else {
-  compatibleDirname = dirname(fileURLToPath(import.meta.url));
-}
-
-const jsCodePath = path.resolve(compatibleDirname, './client.umd.js');
-
-const jsCode = fs.readFileSync(jsCodePath, 'utf-8');
 
 export type HotKey = 'ctrlKey' | 'altKey' | 'metaKey' | 'shiftKey';
 export type CodeOptions = {
@@ -64,139 +35,12 @@ export type CodeOptions = {
   enforcePre?: boolean;
 };
 
-export function getInjectCode(port: number, options?: CodeOptions) {
-  const {
-    hotKeys = ['shiftKey', 'altKey'],
-    showSwitch = false,
-    hideConsole = false,
-    autoToggle = true,
-  } = options || ({} as CodeOptions);
-  return `
-/* eslint-disable */
-if (typeof window !== 'undefined') {
-  if (!document.body.querySelector('code-inspector-component')) {
-    const script = document.createElement('script');
-    script.setAttribute('type', 'text/javascript');
-    script.textContent = ${`${jsCode}`};
-
-    const inspector = document.createElement('code-inspector-component');
-    inspector.port = ${port};
-    inspector.hotkeys = '${(hotKeys ? hotKeys : [])?.join(',')}';
-    inspector.showSwitch = !!${showSwitch};
-    inspector.autoToggle = !!${autoToggle};
-    inspector.hideConsole = !!${hideConsole};
-    document.body.append(inspector);
-  }
-}
-/* eslint-disable */
-`;
-}
-
-// 获取不带文件后缀名的文件路径
-export function getFilenameWithoutExt(filePath: string) {
-  while (path.parse(filePath).ext) {
-    filePath = path.parse(filePath).name;
-  }
-  return filePath;
-}
-
-// 检测是否为符合可注入代码的目标文件
-export function isTargetFile(file: string) {
-  return ['.js', '.ts', '.mjs', '.mts', '.jsx', '.tsx'].some((ext) =>
-    file.endsWith(ext)
-  );
-}
-
-// 检测是否为 nextjs 中的 client 文件
-export function isNextClientFile(code: string) {
-  return (
-    code.trim().startsWith(`"use client"`) ||
-    code.trim().startsWith(`'use client'`) ||
-    code.trim().startsWith(`"use client;"`) ||
-    code.trim().startsWith(`'use client;'`)
-  );
-}
-
-// 检测是否为 useEffect 文件
-export function isUseEffectFile(code: string) {
-  return code.includes(`useEffect(`);
-}
-
-export type RecordInfo = {
-  port: number;
-  entry: string;
-  nextInjectedFile: string;
-  useEffectFile: string;
-  injectAll: boolean;
-};
-
-export async function getServedCode(
-  options: CodeOptions,
-  file: string,
-  code: string,
-  record: RecordInfo
-) {
-  if (!record.port) {
-    // start server
-    record.port = await new Promise((resolve) => {
-      startServer((port) => {
-        resolve(port);
-      }, options?.editor);
-    });
-  }
-
-  if (options?.injectTo) {
-    if (options.injectTo === 'auto') {
-      //
-    } else if (options.injectTo === 'all') {
-      record.injectAll = true;
-    } else if (isAbsolute(options.injectTo)) {
-      if (isTargetFile(record.entry)) {
-        record.entry = getFilenameWithoutExt(options.injectTo);
-      } else {
-        console.error(
-          `The ext of "injectTo" in code-inspector-plugin must in '.js/.ts/.mjs/.mts/.jsx/.tsx'`
-        );
-      }
-    } else {
-      console.error(
-        `"injectTo" in code-inspector-plugin must be 'auto' or 'all' or an absolute file path!`
-      );
-    }
-  }
-  // inject client code to entry file
-  if (!record.entry && isTargetFile(file)) {
-    record.entry = getFilenameWithoutExt(file);
-  }
-  // compatible to nextjs
-  if (
-    !record.nextInjectedFile &&
-    isTargetFile(file) &&
-    isNextClientFile(code)
-  ) {
-    record.nextInjectedFile = getFilenameWithoutExt(file);
-  }
-  // compatible to react ssr but not nextjs
-  if (
-    !record.nextInjectedFile &&
-    !record.useEffectFile &&
-    isTargetFile(file) &&
-    isUseEffectFile(code)
-  ) {
-    record.useEffectFile = getFilenameWithoutExt(file);
-  }
-  if (
-    isTargetFile(file) &&
-    (getFilenameWithoutExt(file) === record.entry ||
-      getFilenameWithoutExt(file) === record.nextInjectedFile ||
-      getFilenameWithoutExt(file) === record.useEffectFile ||
-      record.injectAll)
-  ) {
-    code = `${code}\n${getInjectCode(record.port, {
-      ...(options || {}),
-    })}`;
-  }
-  return code;
-}
-
-export { startServer, enhanceCode, normalizePath, parseSFC };
+export {
+  startServer,
+  enhanceCode,
+  normalizePath,
+  parseSFC,
+  getServedCode,
+} from './server';
+export type { RecordInfo } from './server';
+export { fileURLToPath, isJsTypeFile } from './shared/utils';
