+++ packages/core/types/client/index.d.ts
@@ -68,7 +68,6 @@ export declare class CodeInspectorComponent extends LitElement {
     recordMousePosition: (e: MouseEvent) => void;
     handleMouseUp: () => void;
     switch: (e: Event) => void;
-    eliminateWarning: () => void;
     protected firstUpdated(): void;
     disconnectedCallback(): void;
     render(): import("lit").TemplateResult<1>;
