+++ packages/core/types/shared/constant.d.ts
@@ -4,3 +4,4 @@ export declare const PathName = "data-insp-path";
 export declare const NodeName = "data-insp-node";
 export type Editor = 'atom' | 'code' | 'code_insiders' | 'idea' | 'phpstorm' | 'pycharm' | 'webstorm' | 'hbuilder';
 export declare const DefaultPort = 5678;
+export declare const JsFileExtList: string[];
