+++ packages/core/types/shared/utils.d.ts
@@ -1,0 +1,5 @@
+export declare function fileURLToPath(fileURL: string): string;
+export declare function isJsTypeFile(file: string): boolean;
+export declare function isNextClientFile(code: string): boolean;
+export declare function isUseEffectFile(code: string): boolean;
+export declare function getFilenameWithoutExt(filePath: string): string;
